requirements.txt
confluent-kafka
elasticsearch
pymongo
python-dotenv